Replace wavelet instead of adding as wavelet_en
