+++ apps/wavelet/.gitignore
@@ -1,4 +1,3 @@
-daubechies_x.png
-haar_x.png
-inverse_daubechies_x.png
-inverse_haar_x.png
+build/*
+build_cmake/*
+build_make/*
