+++ apps/wavelet/Makefile
@@ -1,24 +1,46 @@
 include ../support/Makefile.inc
 
-HL_MODULES = daubechies_x.o inverse_daubechies_x.o haar_x.o inverse_haar_x.o halide_runtime.o
+ifdef BUILD_PREFIX
+BUILD_DIR = build_make/$(BUILD_PREFIX)
+else
+BUILD_DIR = build_make
+endif
 
-all: filter
+# If HL_TARGET isn't set, use host
+HL_TARGET ?= host
 
-wavelet: ../../ wavelet.cpp
-	$(MAKE) -C ../../ $(LIB_HALIDE)
-	$(CXX) $(CXXFLAGS) wavelet.cpp -g ../../$(LIB_HALIDE) -o wavelet -lpthread -ldl -lz $(LDFLAGS)
-
-$(HL_MODULES): wavelet
-	./wavelet
-
-filter.o: filter.cpp $(HL_MODULES)
-	$(CXX) $(CXXFLAGS) $(LIBPNG_CXX_FLAGS) -O3 -c filter.cpp
-
-filter: filter.o
-	$(CXX) $(CXXFLAGS) filter.o $(HL_MODULES) $(PNGFLAGS) -o filter -lpthread -ldl
+all: $(BUILD_DIR)/wavelet
 
 clean:
-	rm -f wavelet filter filter.o $(HL_MODULES) $(HL_MODULES:%.o) $(HL_MODULES:%.o=%.h)
+	@rm -rf $(BUILD_DIR)
 
-test: filter
-	./filter ../images/gray.png
+# By default, %.generator is produced by building %_generator.cpp
+$(BUILD_DIR)/%.generator: %_generator.cpp $(GENERATOR_DEPS)
+	@echo Building Generator $(filter %_generator.cpp,$^)
+	@mkdir -p $(BUILD_DIR)
+	@$(CXX) $(CXXFLAGS) -fno-rtti $(filter-out %.h,$^) -lz -o $@
+
+# By default, %.o/.h are produced by executing %.generator
+$(BUILD_DIR)/%.o $(BUILD_DIR)/%.h: $(BUILD_DIR)/%.generator
+	@echo Running Generator $<
+	@mkdir -p $(BUILD_DIR)
+	@$< -g $(notdir $*) -o $(BUILD_DIR) target=$(HL_TARGET)
+
+HL_MODULES = \
+	$(BUILD_DIR)/daubechies_x.o \
+	$(BUILD_DIR)/haar_x.o \
+	$(BUILD_DIR)/inverse_daubechies_x.o \
+	$(BUILD_DIR)/inverse_haar_x.o
+
+$(BUILD_DIR)/wavelet.o: wavelet.cpp $(HL_MODULES)
+	@$(CXX) $(CXXFLAGS) $(LIBPNG_CXX_FLAGS) -I$(BUILD_DIR) -c $< -o $@
+
+$(BUILD_DIR)/wavelet: $(BUILD_DIR)/wavelet.o 
+	@$(CXX) $(CXXFLAGS) $^ $(HL_MODULES) $(PNGFLAGS) -o $@
+
+test: $(BUILD_DIR)/wavelet
+	@echo Testing wavelet...
+	@$< ../images/gray.png $(BUILD_DIR)
+
+# Don't auto-delete the generators.
+.SECONDARY:
