+++ apps/wavelet/README.md
@@ -1,0 +1,1 @@
+wavelet is a trivial app designed to show ahead-of-time Generator usage (with both Make and CMake) instead of direct compilation steps.
