+++ apps/wavelet/daubechies_constants.h
