+++ apps/wavelet/daubechies_x_generator.cpp
