+++ apps/wavelet/haar_x_generator.cpp
