+++ apps/wavelet/inverse_daubechies_x_generator.cpp
