+++ apps/wavelet/inverse_haar_x_generator.cpp
