--- apps/wavelet_gen/.gitignore
@@ -1,3 +1,0 @@
-build/*
-build_cmake/*
-build_make/*
